requirements.txt
streamlit>=1.25.0
torch>=2.0.0
torchvision>=0.15.0
pandas>=2.0.0
Pillow>=9.0.0
numpy
